+++ CHANGES.md
@@ -6,6 +6,7 @@ Change Log
     - Added reprint button to Order print view
     - Updated lite badge price, day
     - Kumo Lite 2018 badge format added
+    - During at con reg, if only one badge type is available auto select it
 
 - Version 3.0.4 (6/18/2018)
     - Show parental consent form checkbox in detail form